Add a `BasePresenter`
In future commits, we’ll have a presenter that inherits from this one.
All presenters will render a base tag with data attributes that contain
information about the content block. Presenters that inherit from the
BasePresenter will be able to override the `content` method and return
the relevant information from the block in the format they need.