Specify gem name
We’ve recently renamed the repo to be more in line with naming
conventions, but not the gem (as this would open up a bit of a can
of worms, as there’s no real support in Rubygems for renaming). As a
result, we need to specify the gem name in the Github actions, as they
default to using the repo name.