symbolise keys in block details
I noticed when working on nested fields,
that for previewing functionalities in
Content Block Manager and Whitehall the
nested fields weren't working, though they were
working in Publishing API when generating
the content for the frontend - for some reason
on the Whitehall side any nested fields in the `details` are returned with the keys as strings.
This prevents that causing a problem hopefully.