Add check for content in fields
Embed codes can be in this format:

`{{embed:{content_type}:{content_id}/{field_path_to_content}}}``

If a user were to accidentally change field path
use an embed code with the wrong fields, we want
to register that in some way to make it easier to
debug.

Here I add Rails logging to catch the instance,
and return the embed code as plain text. As this
Tools repo is used in a couple of places
(Publishing API and Whitehall) it can be quite
difficult to debug when things go wrong without
any logging.

I considered adding Sentry errors but that would
involve adding Sentry to the gem (in case whatever
repo used the gem didn't have Sentry) and having
some kind of error class of our own. This should
be an edge case and easily corrected by the Editor
once they see the embed code not working - if it
does become an issue once in Production we can
look at having more robust handling.