Return field_order from content block
We are introducting a new `field_order` field to the `details` attribute
of a content block, which will allow us to define what order the fields
will appear when rendering an entire block.

This adds a new `field_order` method to the ContentBlock class, which
returns the fields to return and their order. If one is not found, we
default to returning the title, followed by all of the fields in the
`details` object in the order in which we find them.