Add a BlockRenderer
This adds the behaviour we’ll be using for rendering blocks.

Basically, we walk through the provided `field_order` for the
content blocks. If the field is a string, we render it directly
(unless it’s a title). If the field is an object, we walk through
the keys of the object, fetch the block, and attempt to find a
presenter for that block, defaulting to the BasePresenter if we
can’t find it.