Support embedding nested blocks
Before, if someone referenced a block that was embedded in a wider
object (for example) an address embedded within a contact, then the
object would be output as a plain hash (as if it were a string). This
adds functionality to check if the content is a hash, and if it is, try
and find a presenter to use to display the content.