Fetch the `base_tag` from the presenter
At the moment, when attempting to render a telephone number in
isolation, the presenter renders a `div` inside a `span`, which
is not valid HTML. This results in the HTML wrapped inside the
enclosing `span` to be escaped and rendered as text. This is because
when field names are provided, we automatically assume the content
will be wrapped in a `span`

To get around this, we add `BASE_TAG_TYPE`s to the base Field and
Block presenters. If `field_names` are present for an embed code,
we fetch the `BASE_TAG_TYPE` from the presenter and use that.

In most cases, this will be a `span`, but we override this when
rendering a telephone.

I’ve also added a test for rendering an individual telephone, as this
was missing before.