Customise the wrapper class of a block
If the block is being rendered on its own (i.e with a `field_names`)
context, then we want to add an additional `contact` div to wrap the
block in to ensure it styles correctly. We do this by adding a new
`BlockLevelContactItem` module, which we use in all the block level
presenters.

In the next commit, we’ll pass this new argument to the presenters from
the BasePresenter.