Replace special dashes with double/triple dashes
Kramdown (the markdown parses that Govspeak is based on) replaces
double/triple dashes with en/em dashes. This can cause a problem if
an embed code contains double or triple dashes (for example when there
is a name collision for an identifier.

It seems the easiest way to combat this is scan the document for special
dashes, then replace them with double or triple dashes when initializing
the references. This should ensure the codes get found correctly.