Add escape character to regex
At the moment, we’re getting a warning in the logs, which is quite
noisy. Before we added the en and em dashes, we weren’t getting the
warning as the dash was the last character in the sequence. Adding
the escape character ensures the regex still works, but gets rid of
the noise in the logs.