Move render logic to the ContactBlock class
This moves a lot of the presentation logic to the ContentBlock class
itself. If a block is going to be rendered, we fetch the relevant
component and render that way, or we attempt to fetch a field presenter
(defaulting to the base presenter) and render the field that way.

If there is no presenter for a document type, we return the field, which
is the current behaviour.

If no content is found for an embed code, we send a warning to the
logger and rendet the embed code.