Alter how embedded blocks are rendered
Before, we worked on the assumption that all items within the block are
objects, which isn’t the case. This removes the `content_by_block_type`,
which threw a `NameError` when trying to call `.values` on the
`description` string, preferring to loop thorough all the `BLOCK_TYPES`,
then fetching all the blocks for a particular block type.