Tweak dependencies
We’ve had some issues getting the latest version of the Tools repo
because of dependency issues with Govspeak. This makes the versions
of gem dependencies more specific, but looser in some cases. This
matches the approach used in Govspeak here https://github.com/alphagov/govspeak/blob/main/govspeak.gemspec