Switching rendering fn to handle linebreaks.
Currently the address lines are rendered with the content_tag fn
to convert them to renderable html (span) tags.

I've updated this to use the simple_format fn instead as this does
roughly the same thing but will also convert linebreaks into break
tags. As the street_address field is changing to accept multiline
input I think this is needed.