Remove headings from contact blocks
When we first put contact blocks together, we made the decision to
mirror the contacts in the [Contact a department or service team][]
design. We hit a problem when trying to use `h2`s in the blocks, as we
could not guarantee where a block would be added, and this had
accessibility issues.

To get around this, we decided to render contacts within nested
definition lists, reset the margins, padding etc and style the `dt`
elements like `h2` and `h3`s.

HOWEVER, when testing this with users, we found that they expected the
`dt` “headings” for blocks to appear in page navigation, as this is how
headings behave elsewhere.

To get around this issue, we’ve decided to not show the heading of a
block at all, and let the user put in their own headings when embedding
contact blocks.

We do this by changing the `dl` wrapper to be a simple `div` (with the
`vcard` microformat class intact) and removing the `dt` and `dd`s that
wrap the contact(s).

Now when rendering an entire block, we have one definition list, and
when rendering one particular content type, we just render the component
within the wrapping `div`

There are some minor style tweaks to make on the components gem, but
will do this seperately.

[Contact a department or service team]: https://design-system.service.gov.uk/patterns/contact-a-department-or-service-team/